system-engineering/project-α/requirements.txt
# Diana - Phase 1 Dependencies
# LLM Providers
google-genai>=1.0.0

# Vector Database
chromadb>=0.5.0

# Embedding Model
sentence-transformers>=3.0.0

# Folder Watching
watchdog>=4.0.0

# Data Validation
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# PDF Parsing
PyPDF2>=3.0.0

# Configuration
pyyaml>=6.0.0

# Rich CLI
rich>=13.0.0
